create-legacy-oscontainer: use runvm not nested containers
This also changes the push to create a oci-archive that will
be pushed with `cosa push-container-manifest` by the pipeline.
Renamed upload-oscontainer to create-legacy-oscontainer
as the behavior has changed. The argument `--name` is not
required anymore, before it was being used to pass a repository
now it reads the name from the meta.json by default.